+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - reason attribute to PixRequest resource
 - PixFraud resource
 - PixKeyHolmes resource
+- PixInternalTransactionReport resource
 
 ## [0.18.0] - 2026-06-22
 ### Changed